move js file to module
